requirements.txt
# Core Dependencies
fastapi
uvicorn[standard]
python-dotenv
manim
python-multipart

# AI & Cloud Services
google-generativeai
google-cloud-aiplatform
google-auth

# Audio/Visual Processing
librosa
soundfile
pygame
pydub
opencv-python
matplotlib

# Web & Async
aiofiles
httpx
beautifulsoup4

# PDF Processing
pypdf

# Utilities
pathlib
json-logging
typing-extensions
